Menu linked to other activities
I changed a couple of the names of activities to make it more clear what
they were. I added blank activities for each of the buttons on the menu.
Once we have a better idea of what we want the game to look like I can
update the theme of the main menu (logo, title, background, sounds,
etc.) Also, I was wondering about some of the other Java files,
specifically DrawingTheBall, Sweet, and TutorialOne. Are these leftovers
from the video tutorials that can be deleted or are they files we can
repurpose for our project?